+++ packages/algob/sample-project/ts/test/sample-test.ts
@@ -56,7 +56,7 @@ describe("Sample Test", function () {
 		const initialMasterBalance = master.balance();
 
 		try {
-			runtime.executeTx([
+			assert.throws(() => runtime.executeTx([
 				{
 					type: types.TransactionType.TransferAlgo,
 					sign: types.SignType.LogicSignature,
@@ -66,7 +66,7 @@ describe("Sample Test", function () {
 					amountMicroAlgos: amount,
 					payFlags: { totalFee: invalidTxFee },
 				},
-			]);
+			]));
 		} catch (error) {
 			console.log(error);
 		}